+++ packages/adapter-cloudflare/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sveltejs/adapter-cloudflare
 
+## 1.0.0-next.21
+
+### Patch Changes
+
+- only serve `_app/immutable` with immutable cache header, not `_app/version.json` ([#5051](https://github.com/sveltejs/kit/pull/5051))
+
+## 1.0.0-next.20
+
+### Patch Changes
+
+- Add types to pkg.exports ([#5045](https://github.com/sveltejs/kit/pull/5045))
+
 ## 1.0.0-next.19
 
 ### Patch Changes