+++ packages/adapter-vercel/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @sveltejs/adapter-vercel
 
+## 1.0.0-next.56
+
+### Patch Changes
+
+- Use @vercel/nft to include dependencies in lambda without bundling with esbuild, when using v3 build output API ([#4969](https://github.com/sveltejs/kit/pull/4969))
+
+* only serve `_app/immutable` with immutable cache header, not `_app/version.json` ([#5051](https://github.com/sveltejs/kit/pull/5051))
+
+## 1.0.0-next.55
+
+### Patch Changes
+
+- Add types to pkg.exports ([#5045](https://github.com/sveltejs/kit/pull/5045))
+
+## 1.0.0-next.54
+
+### Patch Changes
+
+- [breaking] replace @sveltejs/kit/install-fetch with @sveltejs/kit/node/polyfills ([#4934](https://github.com/sveltejs/kit/pull/4934))
+
+## 1.0.0-next.53
+
+### Patch Changes
+
+- Revert to cjs mode when building for lambda ([#4967](https://github.com/sveltejs/kit/pull/4967))
+
 ## 1.0.0-next.52
 
 ### Patch Changes